[TensorExpr] Implement shape inference for TE.
Since TE operates on a limited subset of ops with a well-defined
semantics, we can easily infer shapes of intermediate and output tensors
given shapes of the inputs.

There is a couple of ops that are not yet supported in the shape
inference, once we add them we could relax the shape info requirements
in the TE fuser: currently it requires all values in the fusion group to
have shapes known and we can change it to only inputs.

ghstack-source-id: cc10d8c63cd9551d33cdc252c43c1a5ceec9f7f6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/41451